requirements.txt
absl-py==0.8.1
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.1
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
defer==1.0.6
dill==0.3.1.1
distro-info===0.18ubuntu0.18.04.1
gast==0.3.2
google-pasta==0.1.8
grpcio==1.25.0
gspan-mining==0.2.2
h5py==2.10.0
httplib2==0.9.2
idna==2.6
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
labelme==3.16.7
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
llvmlite==0.31.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.2.1
mlxtend==0.17.2
multiprocess==0.70.9
netifaces==0.10.4
numba==0.48.0
numpy==1.18.4
oauth==1.0.1
olefile==0.45.1
pandas==1.0.3
pexpect==4.2.1
Pillow==6.2.1
protobuf==3.11.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
PyQt5==5.14.0
PyQt5-sip==12.7.0
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2020.1
pyxdg==0.25
PyYAML==5.2
QtPy==1.9.0
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-learn==0.22.2.post1
scipy==1.4.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.14.0
system-service==0.3
systemd-python==234
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
Werkzeug==0.16.0
wrapt==1.11.2
xkit==0.0.0
zope.interface==4.3.2
